requirements.txt
requests
python-dotenv
openai
tenacity
pandas
openpyxl
pydantic
SQLAlchemy>=2.0
psycopg2-binary
streamlit
gunicorn
alembic
xlsxwriter
slack_sdk
aiohttp
python-dateutil 